# contracts/openapi/order-query.v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Order Query and Administration API",
    "version": "v1",
    "description": "Customers query only their own order history. Administrators can operate fulfillment and create durable Zarinpal refund requests."
  },
  "paths": {
    "/api/v1/orders": {
      "get": {
        "summary": "List the caller's own orders in descending creation order",
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/Limit"}, {"$ref": "#/components/parameters/Cursor"}],
        "responses": {"200": {"description": "Customer order page", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OrderPage"}}}}, "403": {"description": "Customer role required"}, "422": {"description": "Invalid cursor"}}
      }
    },
    "/api/v1/orders/{order_id}": {
      "get": {
        "summary": "Get one order owned by the caller",
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/OrderId"}],
        "responses": {"200": {"description": "Order detail", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Order"}}}}, "404": {"description": "Order not found or not owned"}}
      }
    },
    "/api/v1/orders/admin": {
      "get": {
        "summary": "List all order summaries for administrator review",
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/Limit"}, {"$ref": "#/components/parameters/Cursor"}, {"name": "status", "in": "query", "schema": {"$ref": "#/components/schemas/OrderStatus"}}],
        "responses": {"200": {"description": "Administrator order page", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OrderPage"}}}}, "403": {"description": "Administrator role required"}, "422": {"description": "Invalid cursor or status"}}
      }
    },
    "/api/v1/orders/admin/{order_id}": {
      "get": {
        "summary": "Get operational detail for one order",
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/OrderId"}],
        "responses": {"200": {"description": "Administrator order detail", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AdminOrder"}}}}, "403": {"description": "Administrator role required"}, "404": {"description": "Order not found"}}
      }
    },
    "/api/v1/orders/admin/{order_id}/fulfillment": {
      "patch": {
        "summary": "Advance a paid order through fulfillment and record shipment tracking",
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/OrderId"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/FulfillmentUpdate"}}}},
        "responses": {"200": {"description": "Updated administrator order", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AdminOrder"}}}}, "403": {"description": "Administrator role required"}, "404": {"description": "Order not found"}, "409": {"description": "Invalid fulfillment transition"}, "422": {"description": "Shipment tracking is incomplete"}}
      }
    },
    "/api/v1/orders/admin/{order_id}/refund": {
      "post": {
        "summary": "Create or resume a durable Zarinpal refund request for a confirmed order",
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/OrderId"}, {"$ref": "#/components/parameters/IdempotencyKey"}],
        "responses": {"202": {"description": "Refund request accepted", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RefundRequest"}}}}, "403": {"description": "Administrator role required"}, "404": {"description": "Order not found"}, "409": {"description": "Order is not refundable or the idempotency key conflicts"}}
      }
    }
  },
  "components": {
    "securitySchemes": {"bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}},
    "parameters": {
      "OrderId": {"name": "order_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
      "IdempotencyKey": {"name": "Idempotency-Key", "in": "header", "required": true, "schema": {"type": "string", "minLength": 1, "maxLength": 128}},
      "Limit": {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 20}},
      "Cursor": {"name": "cursor", "in": "query", "schema": {"type": "string", "minLength": 1, "maxLength": 256}}
    },
    "schemas": {
      "OrderStatus": {"enum": ["PENDING", "INVENTORY_RESERVED", "PAYMENT_PENDING", "CONFIRMED", "PROCESSING", "SHIPPED", "DELIVERED", "REFUND_PENDING", "REFUNDED", "CANCELLED"]},
      "OrderSummary": {"type": "object", "required": ["id", "status", "tracking_code", "currency", "total_amount", "created_at", "updated_at"], "properties": {"id": {"type": "string", "format": "uuid"}, "status": {"$ref": "#/components/schemas/OrderStatus"}, "tracking_code": {"type": "string"}, "currency": {"type": "string"}, "total_amount": {"type": "string"}, "created_at": {"type": "string", "format": "date-time"}, "updated_at": {"type": "string", "format": "date-time"}}},
      "OrderPage": {"type": "object", "required": ["items", "next_cursor"], "properties": {"items": {"type": "array", "items": {"$ref": "#/components/schemas/OrderSummary"}}, "next_cursor": {"type": ["string", "null"]}}},
      "OrderItem": {"type": "object", "required": ["variant_id", "sku", "product_name", "unit_amount", "quantity", "attributes"], "properties": {"variant_id": {"type": "string", "format": "uuid"}, "sku": {"type": "string"}, "product_name": {"type": "string"}, "unit_amount": {"type": "string"}, "quantity": {"type": "integer"}, "attributes": {"type": "object", "additionalProperties": {"type": "string"}}}},
      "Fulfillment": {"type": "object", "required": ["carrier", "tracking_number", "updated_at"], "properties": {"carrier": {"type": ["string", "null"]}, "tracking_number": {"type": ["string", "null"]}, "updated_at": {"type": "string", "format": "date-time"}}},
      "FulfillmentUpdate": {"type": "object", "required": ["status"], "properties": {"status": {"enum": ["PROCESSING", "SHIPPED", "DELIVERED"]}, "carrier": {"type": "string", "minLength": 1, "maxLength": 120}, "tracking_number": {"type": "string", "minLength": 1, "maxLength": 160}}},
      "Order": {"allOf": [{"$ref": "#/components/schemas/OrderSummary"}, {"type": "object", "required": ["items", "fulfillment"], "properties": {"items": {"type": "array", "items": {"$ref": "#/components/schemas/OrderItem"}}, "fulfillment": {"oneOf": [{"$ref": "#/components/schemas/Fulfillment"}, {"type": "null"}]}}}]},
      "Transition": {"type": "object", "required": ["from_status", "to_status", "reason", "created_at"], "properties": {"from_status": {"oneOf": [{"$ref": "#/components/schemas/OrderStatus"}, {"type": "null"}]}, "to_status": {"$ref": "#/components/schemas/OrderStatus"}, "reason": {"type": "string"}, "created_at": {"type": "string", "format": "date-time"}}},
      "RefundRequest": {"type": "object", "required": ["order_id", "refund_request_id", "status"], "properties": {"order_id": {"type": "string", "format": "uuid"}, "refund_request_id": {"type": "string", "format": "uuid"}, "status": {"const": "REFUND_PENDING"}}},
      "AdminOrder": {"allOf": [{"$ref": "#/components/schemas/Order"}, {"type": "object", "required": ["customer_id", "customer_email", "delivery_address", "payment_method", "transitions", "invoice", "refund_request_id"], "properties": {"customer_id": {"type": "string", "format": "uuid"}, "customer_email": {"type": ["string", "null"], "format": "email"}, "delivery_address": {"type": "object", "additionalProperties": {"type": "string"}}, "payment_method": {"type": "string"}, "transitions": {"type": "array", "items": {"$ref": "#/components/schemas/Transition"}}, "invoice": {"type": ["object", "null"]}, "refund_request_id": {"type": ["string", "null"], "format": "uuid"}}}]}
    }
  }
}